--- src/main/java/com/baidu/fsg/uid/exception/UidGenerateException.java
@@ -1,75 +1,0 @@
-/*
- * Copyright (c) 2017 Baidu, Inc. All Rights Reserve.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.baidu.fsg.uid.exception;
-
-/**
- * UidGenerateException
- * 
- * @author yutianbao
- */
-public class UidGenerateException extends RuntimeException {
-
-    /**
-     * Serial Version UID
-     */
-    private static final long serialVersionUID = -27048199131316992L;
-
-    /**
-     * Default constructor
-     */
-    public UidGenerateException() {
-        super();
-    }
-
-    /**
-     * Constructor with message & cause
-     * 
-     * @param message
-     * @param cause
-     */
-    public UidGenerateException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    /**
-     * Constructor with message
-     * 
-     * @param message
-     */
-    public UidGenerateException(String message) {
-        super(message);
-    }
-    
-    /**
-     * Constructor with message format
-     * 
-     * @param msgFormat
-     * @param args
-     */
-    public UidGenerateException(String msgFormat, Object... args) {
-        super(String.format(msgFormat, args));
-    }
-
-    /**
-     * Constructor with cause
-     * 
-     * @param cause
-     */
-    public UidGenerateException(Throwable cause) {
-        super(cause);
-    }
-
-}
