--- src/main/java/com/baidu/fsg/uid/utils/ValuedEnum.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2017 Baidu, Inc. All Rights Reserve.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.baidu.fsg.uid.utils;
-
-/**
- * {@code ValuedEnum} defines an enumeration which is bounded to a value, you
- * may implements this interface when you defines such kind of enumeration, that
- * you can use {@link EnumUtils} to simplify parse and valueOf operation.
- *  
- * @author yutianbao
- */
-public interface ValuedEnum<T> {
-    T value();
-}
