--- src/main/java/com/baidu/fsg/uid/worker/entity/WorkerNodeEntity.java
@@ -1,131 +1,0 @@
-/*
- * Copyright (c) 2017 Baidu, Inc. All Rights Reserve.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.baidu.fsg.uid.worker.entity;
-
-import java.util.Date;
-
-import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
-
-import com.baidu.fsg.uid.worker.WorkerNodeType;
-
-/**
- * Entity for M_WORKER_NODE
- *
- * @author yutianbao
- */
-public class WorkerNodeEntity {
-
-    /**
-     * Entity unique id (table unique)
-     */
-    private long id;
-
-    /**
-     * Type of CONTAINER: HostName, ACTUAL : IP.
-     */
-    private String hostName;
-
-    /**
-     * Type of CONTAINER: Port, ACTUAL : Timestamp + Random(0-10000)
-     */
-    private String port;
-
-    /**
-     * type of {@link WorkerNodeType}
-     */
-    private int type;
-
-    /**
-     * Worker launch date, default now
-     */
-    private Date launchDate = new Date();
-
-    /**
-     * Created time
-     */
-    private Date created;
-
-    /**
-     * Last modified
-     */
-    private Date modified;
-
-    /**
-     * Getters & Setters
-     */
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getHostName() {
-        return hostName;
-    }
-
-    public void setHostName(String hostName) {
-        this.hostName = hostName;
-    }
-
-    public String getPort() {
-        return port;
-    }
-
-    public void setPort(String port) {
-        this.port = port;
-    }
-
-    public int getType() {
-        return type;
-    }
-
-    public void setType(int type) {
-        this.type = type;
-    }
-
-    public Date getLaunchDate() {
-        return launchDate;
-    }
-
-    public void setLaunchDateDate(Date launchDate) {
-        this.launchDate = launchDate;
-    }
-
-    public Date getCreated() {
-        return created;
-    }
-
-    public void setCreated(Date created) {
-        this.created = created;
-    }
-
-    public Date getModified() {
-        return modified;
-    }
-
-    public void setModified(Date modified) {
-        this.modified = modified;
-    }
-
-    @Override
-    public String toString() {
-        return ReflectionToStringBuilder.toString(this, ToStringStyle.SHORT_PREFIX_STYLE);
-    }
-
-}
