+++ src/main/java/jp/ne/paypay/uid/buffer/RejectedPutBufferHandler.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2017 Baidu, Inc. All Rights Reserve.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package jp.ne.paypay.uid.buffer;
+
+/**
+ * If tail catches the cursor it means that the ring buffer is full, any more buffer put request will be rejected.
+ * Specify the policy to handle the reject. This is a Lambda supported interface
+ *
+ * @author yutianbao
+ */
+@FunctionalInterface
+public interface RejectedPutBufferHandler {
+
+    /**
+     * Reject put buffer request
+     *
+     * @param ringBuffer
+     * @param uid
+     */
+    void rejectPutBuffer(RingBuffer ringBuffer, long uid);
+}
