+++ src/main/java/jp/ne/paypay/uid/utils/NetUtils.java
@@ -1,0 +1,88 @@
+/*
+ * Copyright (c) 2017 Baidu, Inc. All Rights Reserve.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package jp.ne.paypay.uid.utils;
+
+import java.net.InetAddress;
+import java.net.NetworkInterface;
+import java.net.SocketException;
+import java.util.Enumeration;
+
+/**
+ * NetUtils
+ *
+ * @author yutianbao
+ */
+public final class NetUtils {
+
+    /**
+     * Pre-loaded local address
+     */
+    public static final InetAddress localAddress;
+
+    static {
+        try {
+            localAddress = getLocalInetAddress();
+        } catch (SocketException | IllegalArgumentException e) {
+            throw new IllegalArgumentException("failed to get local ip.", e);
+        }
+    }
+
+    /** Private Constructor */
+    private NetUtils(){}
+
+    /**
+     * Retrieve the first validated local ip address(the Public and LAN ip addresses are validated).
+     *
+     * @return the local address
+     * @throws SocketException the socket exception
+     * @throws IllegalArgumentException no valid local address found
+     */
+    public static InetAddress getLocalInetAddress() throws SocketException {
+        // enumerates all network interfaces
+        Enumeration<NetworkInterface> enu = NetworkInterface.getNetworkInterfaces();
+
+        while (enu.hasMoreElements()) {
+            NetworkInterface ni = enu.nextElement();
+            if (ni.isLoopback()) {
+                continue;
+            }
+
+            Enumeration<InetAddress> addressEnumeration = ni.getInetAddresses();
+            while (addressEnumeration.hasMoreElements()) {
+                InetAddress address = addressEnumeration.nextElement();
+
+                // ignores all invalidated addresses
+                if (address.isLinkLocalAddress() || address.isLoopbackAddress() || address.isAnyLocalAddress()) {
+                    continue;
+                }
+
+                return address;
+            }
+        }
+
+        throw new IllegalArgumentException("No validated local address!");
+    }
+
+    /**
+     * Retrieve local address
+     *
+     * @return the string local address
+     */
+    public static String getLocalAddress() {
+        return localAddress.getHostAddress();
+    }
+
+}
