--- src/test/java/com/baidu/fsg/uid/CachedUidGeneratorTest.java
@@ -1,128 +1,0 @@
-package com.baidu.fsg.uid;
-
-import com.baidu.fsg.uid.impl.CachedUidGenerator;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import javax.annotation.Resource;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import java.util.concurrent.ConcurrentSkipListSet;
-import java.util.concurrent.atomic.AtomicInteger;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-/**
- * Test for {@link CachedUidGenerator}
- *
- * @author yutianbao
- */
-@ExtendWith(SpringExtension.class)
-@ContextConfiguration(locations = { "classpath:uid/cached-uid-spring.xml" })
-public class CachedUidGeneratorTest {
-    private static final int SIZE = 7000000; // 700w
-    private static final boolean VERBOSE = false;
-    private static final int THREADS = Runtime.getRuntime().availableProcessors() << 1;
-
-    @Resource
-    private UidGenerator uidGenerator;
-
-    /**
-     * Test for serially generate
-     *
-     * @throws IOException
-     */
-    @Test
-    public void testSerialGenerate() throws IOException {
-        // Generate UID serially
-        Set<Long> uidSet = new HashSet<>(SIZE);
-        for (int i = 0; i < SIZE; i++) {
-            doGenerate(uidSet, i);
-        }
-
-        // Check UIDs are all unique
-        checkUniqueID(uidSet);
-    }
-
-    /**
-     * Test for parallel generate
-     *
-     * @throws InterruptedException
-     * @throws IOException
-     */
-    @Test
-    public void testParallelGenerate() throws InterruptedException, IOException {
-        AtomicInteger control = new AtomicInteger(-1);
-        Set<Long> uidSet = new ConcurrentSkipListSet<>();
-
-        // Initialize threads
-        List<Thread> threadList = new ArrayList<>(THREADS);
-        for (int i = 0; i < THREADS; i++) {
-            Thread thread = new Thread(() -> workerRun(uidSet, control));
-            thread.setName("UID-generator-" + i);
-
-            threadList.add(thread);
-            thread.start();
-        }
-
-        // Wait for worker done
-        for (Thread thread : threadList) {
-            thread.join();
-        }
-
-        // Check generate 700w times
-        Assertions.assertEquals(SIZE, control.get());
-
-        // Check UIDs are all unique
-        checkUniqueID(uidSet);
-    }
-
-    /**
-     * Worker run
-     */
-    private void workerRun(Set<Long> uidSet, AtomicInteger control) {
-        for (;;) {
-            int myPosition = control.updateAndGet(old -> old == SIZE ? SIZE : old + 1);
-            if (myPosition == SIZE) {
-                return;
-            }
-
-            doGenerate(uidSet, myPosition);
-        }
-    }
-
-    /**
-     * Do generating
-     */
-    private void doGenerate(Set<Long> uidSet, int index) {
-        long uid = uidGenerator.getUID();
-        String parsedInfo = uidGenerator.parseUID(uid);
-        boolean existed = !uidSet.add(uid);
-        if (existed) {
-            System.out.println("Found duplicate UID " + uid);
-        }
-
-        // Check UID is positive, and can be parsed
-        Assertions.assertTrue(uid > 0L);
-        Assertions.assertTrue(StringUtils.isNotBlank(parsedInfo));
-
-        if (VERBOSE) {
-            System.out.println(Thread.currentThread().getName() + " No." + index + " >>> " + parsedInfo);
-        }
-    }
-
-    /**
-     * Check UIDs are all unique
-     */
-    private void checkUniqueID(Set<Long> uidSet) throws IOException {
-        System.out.println(uidSet.size());
-        Assertions.assertEquals(SIZE, uidSet.size());
-    }
-
-}
